+++ ullm/zero_one.py
@@ -6,8 +6,8 @@ from .base import (
     RemoteLanguageModel,
     RemoteLanguageModelMetaInfo,
 )
-from .openai import (
-    OpenAICompatibleModel,
+from .openai import OpenAICompatibleModel
+from .openai_types import (
     OpenAIRequestBody,
 )
